feat(modularização) refatora abertura e fechamento de modais usando data-attributes e event listeners
- Remove onclick inline do HTML para modais
- Implementa funções JS para gerenciar abertura e fechamento via data-modal e data-close-modal
- Reaproveita código para adicionar event listeners dinamicamente
